requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
curl_cffi==0.7.1
frozendict==2.4.4
idna==3.7
multitasking==0.0.11
numpy==1.26.4
pandas==2.2.2
peewee==3.17.5
platformdirs==4.2.2
protobuf==5.27.1
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websockets==10.4
yfinance==0.2.40
ccxt==4.4.6
matplotlib==3.9.2
scipy==1.14.1
python-dotenv==1.0.1